Consolidate approve button
